+++ src/lib/Errors.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2021 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2023 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -13,143 +13,263 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+/**
+ * The error class for Clap.
+ */
 
-import * as $Exceptions from '@litert/exception';
+export abstract class ClapError extends Error {
 
-export const errorRegistry = $Exceptions.createExceptionRegistry({
-    'module': 'clap.litert.org',
-    'types': {
-        'prepare': {
-            'index': $Exceptions.createIncreaseCodeIndex(1)
-        },
-        'runtime': {
-            'index': $Exceptions.createIncreaseCodeIndex(1000)
-        }
+    public constructor(
+        /**
+         * The name of the error.
+         */
+        name: string,
+        /**
+         * The message of the error.
+         */
+        message: string,
+        public readonly ctx: Record<string, unknown> = {},
+        /**
+         * The metadata of the error.
+         */
+        public readonly origin: unknown = null
+    ) {
+
+        super(message);
+        this.name = name;
     }
-});
+}
 
-export const E_NO_SUCH_COMMAND = errorRegistry.register({
-    name: 'no_such_command',
-    message: 'The command does not exist.',
-    metadata: {},
-    type: 'prepare'
-});
+export const E_NO_SUCH_COMMAND = class extends ClapError {
 
-export const E_NO_SUCH_OPTION = errorRegistry.register({
-    name: 'no_such_option',
-    message: 'The option does not exist.',
-    metadata: {},
-    type: 'prepare'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
 
-export const E_NO_SUCH_FLAG = errorRegistry.register({
-    name: 'no_such_flag',
-    message: 'The flag does not exist.',
-    metadata: {},
-    type: 'prepare'
-});
+        super(
+            'no_such_command',
+            'The command does not exist.',
+            ctx,
+            origin,
+        );
+    }
+};
 
-export const E_DUP_OPTION_NAME = errorRegistry.register({
-    name: 'dup_option_name',
-    message: 'The name of option has already been used.',
-    metadata: {},
-    type: 'prepare'
-});
+export const E_NO_SUCH_OPTION = class extends ClapError {
 
-export const E_DUP_COMMAND_NAME = errorRegistry.register({
-    name: 'dup_command_name',
-    message: 'The name of command has already been used.',
-    metadata: {},
-    type: 'prepare'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
 
-export const E_DUP_FLAG_NAME = errorRegistry.register({
-    name: 'dup_flag_name',
-    message: 'The name of flag has already been used.',
-    metadata: {},
-    type: 'prepare'
-});
+        super(
+            'no_such_option',
+            'The option does not exist.',
+            ctx,
+            origin,
+        );
+    }
+};
 
-export const E_DUP_OPTION_SHORTCUT = errorRegistry.register({
-    name: 'dup_option_shortcut',
-    message: 'The shortcut of option has already been used.',
-    metadata: {},
-    type: 'prepare'
-});
+export const E_NO_SUCH_FLAG = class extends ClapError {
 
-export const E_DUP_COMMAND_SHORTCUT = errorRegistry.register({
-    name: 'dup_command_shortcut',
-    message: 'The shortcut of command has already been used.',
-    metadata: {},
-    type: 'prepare'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
 
-export const E_DUP_FLAG_SHORTCUT = errorRegistry.register({
-    name: 'dup_flag_shortcut',
-    message: 'The shortcut of flag has already been used.',
-    metadata: {},
-    type: 'prepare'
-});
+        super(
+            'no_such_flag',
+            'The flag does not exist.',
+            ctx,
+            origin,
+        );
+    }
+};
 
-export const E_INVALID_OPTION_NAME = errorRegistry.register({
-    name: 'invalid_option_name',
-    message: 'The name of option is malformed.',
-    metadata: {},
-    type: 'prepare'
-});
+export const E_DUP_OPTION_NAME = class extends ClapError {
 
-export const E_INVALID_COMMAND_NAME = errorRegistry.register({
-    name: 'invalid_command_name',
-    message: 'The name of command is malformed.',
-    metadata: {},
-    type: 'prepare'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
 
-export const E_INVALID_FLAG_NAME = errorRegistry.register({
-    name: 'invalid_flag_name',
-    message: 'The name of flag is malformed.',
-    metadata: {},
-    type: 'prepare'
-});
+        super(
+            'dup_option_name',
+            'The name of option has already been used.',
+            ctx,
+            origin,
+        );
+    }
+};
 
-export const E_INVALID_OPTION_SHORTCUT = errorRegistry.register({
-    name: 'invalid_option_shortcut',
-    message: 'The shortcut of option is malformed.',
-    metadata: {},
-    type: 'prepare'
-});
+export const E_DUP_COMMAND_NAME = class extends ClapError {
 
-export const E_INVALID_COMMAND_SHORTCUT = errorRegistry.register({
-    name: 'invalid_command_shortcut',
-    message: 'The shortcut of command is malformed.',
-    metadata: {},
-    type: 'prepare'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
 
-export const E_INVALID_FLAG_SHORTCUT = errorRegistry.register({
-    name: 'invalid_flag_shortcut',
-    message: 'The shortcut of flag is malformed.',
-    metadata: {},
-    type: 'prepare'
-});
+        super(
+            'dup_command_name',
+            'The name of command has already been used.',
+            ctx,
+            origin,
+        );
+    }
+};
 
-export const E_OPTION_VALUE_REQUIRED = errorRegistry.register({
-    name: 'option_value_required',
-    message: 'The value of option is required.',
-    metadata: {},
-    type: 'runtime'
-});
+export const E_DUP_FLAG_NAME = class extends ClapError {
 
-export const E_COMMAND_REQUIRED = errorRegistry.register({
-    name: 'command_required',
-    message: 'The command is required before arguments.',
-    metadata: {},
-    type: 'runtime'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
 
-export const E_ARGUMENTS_LACKED = errorRegistry.register({
-    name: 'arguments_lacked',
-    message: 'Too few arguments.',
-    metadata: {},
-    type: 'runtime'
-});
+        super(
+            'dup_flag_name',
+            'The name of flag has already been used.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_DUP_OPTION_SHORTCUT = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'dup_option_shortcut',
+            'The shortcut of option has already been used.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_DUP_COMMAND_SHORTCUT = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'dup_command_shortcut',
+            'The shortcut of command has already been used.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_DUP_FLAG_SHORTCUT = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'dup_flag_shortcut',
+            'The shortcut of flag has already been used.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_INVALID_OPTION_NAME = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'invalid_option_name',
+            'The name of option is malformed.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_INVALID_COMMAND_NAME = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'invalid_command_name',
+            'The name of command is malformed.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_INVALID_FLAG_NAME = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'invalid_flag_name',
+            'The name of flag is malformed.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_INVALID_OPTION_SHORTCUT = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'invalid_option_shortcut',
+            'The shortcut of option is malformed.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_INVALID_COMMAND_SHORTCUT = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'invalid_command_shortcut',
+            'The shortcut of command is malformed.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_INVALID_FLAG_SHORTCUT = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'invalid_flag_shortcut',
+            'The shortcut of flag is malformed.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_OPTION_VALUE_REQUIRED = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'option_value_required',
+            'The value of option is required.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_COMMAND_REQUIRED = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'command_required',
+            'The command is required before arguments.',
+            ctx,
+            origin,
+        );
+    }
+};
+
+export const E_ARGUMENTS_LACKED = class extends ClapError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin?: unknown) {
+
+        super(
+            'arguments_lacked',
+            'Too few arguments.',
+            ctx,
+            origin,
+        );
+    }
+};