+++ src/lib/Utils.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2021 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2023 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -75,7 +75,7 @@ export function isFullAssignExpr(token: string): boolean {
 
 export function extractFullAssignExpr(expr: string): string[] {
 
-    const segs = expr.split('=');
+    const pieces = expr.split('=');
 
-    return [segs[0].slice(2), segs.slice(1).join('=')];
+    return [pieces[0].slice(2), pieces.slice(1).join('=')];
 }